requirements.txt
# MedRiskEval Requirements
# Core dependencies
datasets>=2.0.0
pydantic>=2.0.0
typer>=0.9.0
pyyaml>=6.0.0
openai>=1.0.0
httpx>=0.24.0
rich>=13.0.0
tqdm>=4.0.0

# Optional: vLLM for local model serving
# vllm>=0.2.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
